pyproject.toml
[project]
name = "archer"
version = "0.1.0"
description = "A system for validating secrets using YAML templates"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "jsonpath-ng>=1.5.0",
    "click>=8.0.0",
]
readme = "README.md"
requires-python = ">=3.8"

[project.scripts]
archer = "archer.main:validate_secret"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
lint.select = ["W293"]